Refactor import paths and introduce new providers for improved modularity
- Updated import paths for BattleClock and BattleRoundPhase providers to use absolute paths for consistency.
- Added BattleClockProvider and BattleRoundPhaseProvider to manage game state and phases effectively.
- Enhanced layout component imports for better organization and maintainability.
