+++ apps/web/src/app/layout.tsx
@@ -2,12 +2,12 @@ import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "@workspace/shadcn/globals.css";
 import "@mysten/dapp-kit/dist/index.css";
-import { Providers } from "@/app/providers/Providers";
 import AppBar from "@/components/AppBar";
 import { ConfettiEffect } from "@/components/ConfettiEffect";
 // import { ChallengeOverlay } from "@/components/ChallengeOverlay";
 import { ResultView } from "@/components/ResultView";
 import { mockCoinMetadata } from "@/mock/mockData";
+import { Providers } from "@/providers/Providers";
 import { Toaster } from "sonner";
 
 export const runtime = "edge";
