+++ apps/web/src/components/ResultView.tsx
@@ -1,11 +1,11 @@
 "use client";
-import { useBattleClock } from "@/app/providers/BattleClockProvider";
 import DominanceRechart from "@/components/DominanceRechart";
 import type {
   ChartDataPoint,
   PreparedCoinMeta,
 } from "@/components/DominanceRechart";
 import { mockCoinMetadata, mockDominanceChartData } from "@/mock/mockData";
+import { useBattleClock } from "@/providers/BattleClockProvider";
 import type { RoundCoin } from "@/types/roundcoin";
 import { AnimatePresence, motion } from "framer-motion";
 import { Globe, Send, Twitter } from "lucide-react";
